services/requirements.txt
psycopg2-binary
dj-database-url
django
djangorestframework
django-cors-headers
